backend/requirements.txt
# ── Core framework ─────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ── Configuration ──────────────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Data processing ────────────────────────────────────────────────
pandas>=2.1.0
openpyxl>=3.1.0

# ── PDF parsing ────────────────────────────────────────────────────
pdfplumber>=0.10.0
pymupdf>=1.23.0        # PyMuPDF — reliable image extraction from PDFs

# ── Vector database ────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── LLM & Embedding providers ─────────────────────────────────────
openai>=1.10.0
anthropic>=0.18.0

# ── Local embeddings + cross-encoder reranker ─────────────────────
sentence-transformers>=2.3.0

# ── BM25 keyword search ────────────────────────────────────────────
rank-bm25>=0.2.2

# ── Async PostgreSQL (asyncpg — raw driver, no ORM) ───────────────
asyncpg>=0.29.0

# ── Ollama Python SDK ─────────────────────────────────────────────
ollama>=0.4.0

# ── Vision image processing ────────────────────────────────────────
pillow>=10.0.0

# ── Web search & charts ────────────────────────────────────────────
beautifulsoup4>=4.12.0
httpx>=0.27.0
yfinance>=0.2.40
lxml>=5.0.0
duckduckgo-search>=6.2.0

# ── Authentication ──────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
